+++ CHANGELOG.md
@@ -20,6 +20,7 @@ Journal des modifications
 - Formulaire de création d'un individu à partir d'un utilisateur : initialisation manquante du nom patronymique.
 - Module StepStar : mise à jour du script bash d'envoi quotidien des thèses vers STEP/STAR.
 - Module StepStar : plus d'envoi de la date d'abandon (recommandation de l'ABES).
+- Module StepStar : ajout dans le TEF des IdRef de tous les acteurs de la thèse (doctorant, directeur, codirecteur, etc.)
 - Suppression du menu 'Administration > Compléments d'individu' devenu inutile.
 - Amélioration visuelle des pages de création/modification de structure (Etab/ED/UR).
 - Affichage des co-encadrants sur l'annuraire des thèses (au survol de la souris sur la direction de thèse)